contributors/adobe/submitted/css-compositing/mix-blend-mode/mix-blend-mode-parsing.html

<!DOCTYPE html>
<html>
	<head>
		<title>CSS Test: CSS Blending - parsing mix-blend-mode property</title>
		<link rel="author" title="MIRELA BUDAES" href="mbudaes@adobe.com">
		<link rel="help" href="http://dev.w3.org/fxtf/compositing-1/#mix-blend-mode">
		<meta name="flags" content="dom">
		<meta name="assert" content="Value specified for mix-blend-mode property should be parsed correctly">
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<link rel="stylesheet" href="/resource/testharness.css">
	</head>
	<body>
		<div id="log"></div>
	</body>
		<script type="text/javascript">
            function testParse(declaration) {
                var div = document.createElement("div");
                div.setAttribute("style", declaration);
                return div.style.mixBlendMode;
            }

            var blendModes = ["normal", "multiply", "screen", "overlay", "darken", "lighten", "color-dodge", "color-burn",
                      "hard-light", "soft-light", "difference", "exclusion", "hue", "saturation", "color", "luminosity"];

            for (var i = 0; i < blendModes.length; i++) {

            	test(function() {assert_equals(testParse("mix-blend-mode: " + blendModes[i]), blendModes[i] )}, "Mix-blend-mode " + blendModes[i]);
            };

		</script>
</html>
